Refactor and break out code
